requirements.txt
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
requests>=2.25.0
tenacity>=8.0.0
tqdm>=4.60.0